requirements.txt
dashscope==1.15.0
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.1.8
matplotlib==3.8.3
numpy==1.26.4
openai==1.30.4
pandas==1.5.3
pillow==10.2.0
pygwalker==0.4.8.7
pywebio==1.8.3
sqlalchemy==2.0.29
pyyaml==6.0.1